% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/26_jalalidate_other_functions.R
\name{change_date_separator}
\alias{change_date_separator}
\title{change separator}
\usage{
change_date_separator(date_char, new_separator, ...)
}
\arguments{
\item{date_char}{Jalali date (character)}

\item{new_separator}{character (valid separator)}

\item{...}{passing options ("MIN_YEAR" and/or "MAX_YEAR") to override package options (\link{jdopt_get_options})}
}
\value{
Jalali date character
}
\description{
change the Jalali date character separators. For invalid inputs, returns input without change
}
\examples{
change_date_separator(c("1350/01/02", "14021220", "1402/1/40"), "+", VALID_SEPARATORS=c("+"))
# [1] "1350+01+02" "1402+12+20" "1402/1/40"
}
