% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/24_jalalidate_generic_functions.R
\name{diffdate}
\alias{diffdate}
\title{distance of two JalaliDate}
\usage{
diffdate(x, y)
}
\arguments{
\item{x}{JalaliDate object}

\item{y}{JalaliDate object}
}
\value{
double
}
\description{
calculate distance of two JalaliDate, that is, subtracts values of two JalaliDate and
return
}
\examples{
x <- JalaliDate(c(1, 2))
y <- JalaliDate(c(10, 12, 13, 20, 50))
diffdate(x,y)
# [1]  -9 -10  NA  NA  NA
# Warning message:
#  In diff.JalaliDate(x, y) : The length of two vectors aren't equal!
}
