% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/26_jalalidate_other_functions.R
\name{is_jalali_leap_year}
\alias{is_jalali_leap_year}
\title{check Jalali leap year}
\usage{
is_jalali_leap_year(year)
}
\arguments{
\item{year}{double}
}
\value{
logical
}
\description{
check if a Jalali year is leap year
}
\details{
for details of calculation see \url{https://learn.microsoft.com/en-us/dotnet/fundamentals/runtime-libraries/system-globalization-persiancalendar}
}
\examples{
is_jalali_leap_year(c(1402, 1403, 1404))
# [1] FALSE  TRUE FALSE

}
