% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/26_jalalidate_other_functions.R
\name{is_valid_separator}
\alias{is_valid_separator}
\title{check separator}
\usage{
is_valid_separator(separator)
}
\arguments{
\item{separator}{character}
}
\value{
list of validation result and related message
}
\description{
Checks whether a separator can be among the set of valid separators
}
\examples{
is_valid_separator("+")
# $result
# [1] TRUE
#
# $message
# [1] ""

is_valid_separator("+/")
# $result
# [1] FALSE
#
# $message
# [1] "The number of character of the separator must be 0 or 1!"
}
