% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/26_jalalidate_other_functions.R
\name{jalali_year_weeks}
\alias{jalali_year_weeks}
\title{list of a year weeks}
\usage{
jalali_year_weeks(year, ...)
}
\arguments{
\item{year}{double}

\item{...}{passing options ("MIN_YEAR" and/or "MAX_YEAR") to override package options (\link{jdopt_get_options})}
}
\value{
data.frame
}
\description{
create a data.frame consists start and end days of a year weeks
}
\examples{
jalali_year_weeks(1402)
 #    week     f      l
 #1     1 1402/01/01 1402/01/04
 #2     2 1402/01/05 1402/01/11
 #..............................
 #52   52 1402/12/19 1402/12/25
 #53   53 1402/12/26 1402/12/29
}
