% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/10_jdate_options.R
\name{jdopt_reset}
\alias{jdopt_reset}
\title{reset options to initial values}
\usage{
jdopt_reset()
}
\value{
options list that includes DEFAULT_SEPARATOR, VALID_SEPARATORS, MIN_YEAR, MAX_YEAR
}
\description{
return options value to factory settings
}
\details{
The initial values, or factory settings, are: DEFAULT_SEPARATOR = "/"
, VALID_SEPARATORS = c("", "-", "/"), MIN_YEAR = 1200L, and MAX_YEAR = 1500L
}
\examples{
res <- jdopt_reset()
res

#$DEFAULT_SEPARATOR
#[1] "/"
#
#$VALID_SEPARATORS
#[1] ""  "-" "/"
#
#$MIN_YEAR
#[1] 1200
#
#$MAX_YEAR
#[1] 1500

}
