% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/10_jdate_options.R
\name{jdopt_set_min_max_year}
\alias{jdopt_set_min_max_year}
\title{setting range of valid year}
\usage{
jdopt_set_min_max_year(min_year, max_year)
}
\arguments{
\item{min_year}{integer}

\item{max_year}{integer}
}
\value{
options list that includes DEFAULT_SEPARATOR, VALID_SEPARATORS, MIN_YEAR, MAX_YEAR
}
\description{
Determining the minimum and maximum valid value of Jalali date year
}
\details{
'min_year' must be equal or lower than 'max_year' and both must be integer.
Minimum value of 'min_year' is 0 and maximum value of 'max_year' is 9999.
}
\examples{
JalaliDate(list(1000,1,1))
# [1] NA
jdopt_set_min_max_year(100L, 2000L)
JalaliDate(list(1000,1,1))
# [1] "1000/01/01"

}
