% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/24_jalalidate_generic_functions.R
\name{Operators}
\alias{Operators}
\title{Operators}
\usage{
Operators(x, y)
}
\arguments{
\item{x}{JalaliDate or numeric}

\item{y}{JalaliDate or numeric}
}
\value{
JalaliDate
}
\description{
Perform arithmetic operations
}
\details{
Only +, - operators work with JalaliDate objects in some cases. If each of the two
arguments are JalaliDate, the '-' operator calculates the distance of two dates (see \link{diffdate}).
}
\examples{
JalaliDate("1395/10/11") + 1
# [1] "1395/10/12"
JalaliDate("1403/08/10") - 367
# [1] "1402/08/08"
JalaliDate("1403/09/10") - JalaliDate("1403/08/10")
# [1] 30
}
