% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/24_jalalidate_generic_functions.R
\name{print.JalaliDate}
\alias{print.JalaliDate}
\title{print}
\usage{
\method{print}{JalaliDate}(x, ...)
}
\arguments{
\item{x}{JalaliDate}

\item{...}{for future usages}
}
\value{
display
}
\description{
print
}
\examples{
print(JalaliDate(1))
# [1] "1375/01/02"
}
