% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RJalaliDate-package.R
\docType{package}
\name{RJalaliDate-package}
\alias{RJalaliDate}
\alias{RJalaliDate-package}
\title{RJalaliDate: Handling Jalali Date (Persian / Solar Hijri)}
\description{
Jalali calendar, or solar Hijri, is calendar of Iran and Afghanistan (\url{https://en.wikipedia.org/wiki/Solar_Hijri_calendar}). This package is designed to working with Jalali date. For this purpose, It defines JalaliDate class that is similar to Date class.
}
\author{
\strong{Maintainer}: Hosein Rabiee \email{hosein.rabiee@hotmail.com} (\href{https://orcid.org/0000-0002-9379-5476}{ORCID}) [copyright holder]

}
\keyword{internal}
