% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/24_jalalidate_generic_functions.R
\name{today.JalaliDate}
\alias{today.JalaliDate}
\title{today as Jalali}
\usage{
today.JalaliDate()
}
\value{
JalaliDate object
}
\description{
return JalaliDate object of today
}
\examples{
today.JalaliDate()
# [1] "1403/04/31"
}
