% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/24_jalalidate_generic_functions.R
\name{weekdays.JalaliDate}
\alias{weekdays.JalaliDate}
\title{day of week}
\usage{
\method{weekdays}{JalaliDate}(x, abbreviate = NULL)
}
\arguments{
\item{x}{JalaliDate object}

\item{abbreviate}{not applicable in Persian language}
}
\value{
character
}
\description{
return the day of week in Persian
#'
}
\examples{
weekdays(JalaliDate(1))
#[1] `r stringi::stri_unescape_unicode("\\u067E\\u0646\\u062C\\u0020\\u0634\\u0646\\u0628\\u0647")`
}
