% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/24_jalalidate_generic_functions.R
\name{yearweek}
\alias{yearweek}
\title{number of week}
\usage{
yearweek(x, ...)
}
\arguments{
\item{x}{JalaliDate object}

\item{...}{for future usage}
}
\value{
list of current and last week number and label
}
\description{
It shows which week of the year the desired date is.
}
\examples{
yearweek(JalaliDate(321))
# $week_number
# [1] "47"
#
# $week_label
# [1] "75W47"
#
# $last_week_number
# [1] "46"
#
# $last_week_label
# [1] "75W46"
}
